requirements.txt
PyGithub
python-dotenv
reportlab
